refactor: change the repository directory path
